+++ quilt-sync/ui/src/components/buttons/certify_latest.rs
@@ -7,7 +7,7 @@ use super::IconButton;
 pub fn CertifyLatest(on_click: impl Fn(leptos::ev::MouseEvent) + 'static) -> impl IntoView {
     view! {
         <IconButton on_click=UnsyncCallback::new(on_click)>
-            "Promote my commit"
+            "Promote my revision"
         </IconButton>
     }
 }